Reordered navbar, made login widget vanish and MainDiv scroll to top on navigation
